sakha-chatbot/requirements.txt
flask==2.0.1
google-generativeai
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.21.0
python-dotenv==0.19.0
gunicorn==20.1.0 